requirements.txt

addict==2.4.0
cachetools==4.2.1
cupy-cuda102==9.0.0
cycler==0.10.0
Cython==0.29.23
einops==0.3.0
h5py==3.2.1
numpy==1.20.1
opencv-python==4.5.1.48
Pillow==8.1.2
protobuf==3.15.6
PyYAML==5.4.1
scipy==1.6.3
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
timm
torch==1.8.1
torchvision==0.9.1
tqdm==4.59.0

